--- src/docs/asciidoc/friend/FriendChangeStatus.adoc
@@ -1,4 +1,0 @@
-[[friendship-change-status]]
-=== 친구 요청 승낙 / 거부
-
-operation::friend-change-status[snippets='http-request,path-parameters,request-fields,http-response']
