--- src/main/kotlin/com/beside/groubing/groubingserver/domain/blocked/exception/BlockedMemberInputException.kt
@@ -1,5 +1,0 @@
-package com.beside.groubing.groubingserver.domain.blocked.exception
-
-class BlockedMemberInputException(
-    override val message: String
-) : RuntimeException("[BlockedMemberInputException] $message")
