+++ src/main/kotlin/com/beside/groubing/groubingserver/domain/blockedmember/exception/BlockedMemberInputException.kt
@@ -1,0 +1,5 @@
+package com.beside.groubing.groubingserver.domain.blockedmember.exception
+
+class BlockedMemberInputException(
+    override val message: String
+) : RuntimeException("[BlockedMemberInputException] $message")
