+++ src/main/kotlin/com/beside/groubing/groubingserver/domain/blockedmember/payload/request/BlockingMemberRequest.kt
@@ -1,0 +1,10 @@
+package com.beside.groubing.groubingserver.domain.blockedmember.payload.request
+
+import jakarta.validation.constraints.Min
+import jakarta.validation.constraints.NotNull
+
+data class BlockingMemberRequest(
+    @field:NotNull(message = "회원 아이디를 입력해 주세요.")
+    @field:Min(value = 1L, message = "유효하지 않은 회원 아이디입니다.")
+    val targetMemberId: Long
+)
