+++ src/main/kotlin/com/beside/groubing/groubingserver/domain/blockedmember/payload/response/BlockedMemberResponse.kt
@@ -1,0 +1,12 @@
+package com.beside.groubing.groubingserver.domain.blockedmember.payload.response
+
+import com.beside.groubing.groubingserver.domain.member.domain.Member
+
+data class BlockedMemberResponse(
+    val id: Long,
+    val email: String,
+    val nickname: String,
+    val profileUrl: String?
+) {
+    constructor(friend: Member) : this(friend.id, friend.email, friend.nickname, friend.profile?.url)
+}
