+++ src/main/kotlin/com/beside/groubing/groubingserver/domain/friend/application/FriendAcceptService.kt
@@ -1,7 +1,6 @@
 package com.beside.groubing.groubingserver.domain.friend.application
 
 import com.beside.groubing.groubingserver.domain.friend.dao.FriendFindDao
-import com.beside.groubing.groubingserver.domain.friend.domain.FriendRepository
 import com.beside.groubing.groubingserver.domain.friend.domain.FriendStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -9,12 +8,10 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional
 class FriendAcceptService(
-    private val friendRepository: FriendRepository,
     private val friendFindDao: FriendFindDao
 ) {
     fun accept(id: Long) {
         val friendship = friendFindDao.findById(id)
         friendship.status = FriendStatus.ACCEPT
-        friendRepository.save(friendship)
     }
 }
